+++ docs/examples/_setup.jl
@@ -137,7 +137,7 @@ if !@isdefined(_BVD_SETUP_LOADED)
     frozen_C(c) = vec(Array(frozen_by_cutoff[c].chn[:C_T]))
     if RUN_SENSITIVITY
         chn_joint_community_delay = _fits["sens_community_delay"]
-        chn_joint_fast_clock = _fits["sens_fast_clock"]
+        chn_joint_exp_growth_clock = _fits["sens_exp_growth_clock"]
     end
 
     posterior_C_joint = vec(Array(chn_joint[:C_T]));